Field: Move dispatcher to Form
